perf: unsigned fastAbs

+++ firmware/src/BSP/A4950.c
@@ -301,8 +301,8 @@ void phase_voltage_command(int16_t U_a, int16_t U_b, uint16_t curr_lim)
 	set_curr(curr_lim, curr_lim); 
 
 	uint16_t U_in = GetMotorVoltage_mV();
-	uint16_t duty_a = (uint16_t) ((uint32_t)fastAbs(U_a) * PWM_TIM_MAX / U_in);
-	uint16_t duty_b = (uint16_t) ((uint32_t)fastAbs(U_b) * PWM_TIM_MAX / U_in);
+	uint16_t duty_a = (uint16_t)(fastAbs(U_a) * PWM_TIM_MAX / U_in);
+	uint16_t duty_b = (uint16_t)(fastAbs(U_b) * PWM_TIM_MAX / U_in);
 	setPWM_bridgeA(duty_a, (U_a > 0)); //PWM12
 	setPWM_bridgeB(duty_b, liveMotorParams.swapPhase ? (U_b < 0) : (U_b > 0)); //PWM34
 }
